fix: home/rewards views

+++ apps-receipt-rewards/src/main/java/com/mytiki/apps_receipt_rewards/more/MoreContributor.kt
@@ -2,5 +2,5 @@ package com.mytiki.apps_receipt_rewards.more
 
 data class MoreContributor(
     val name: String,
-    val percentage: Float
+    val percentage: Double
 )
